fix(migrations): 🐛 Add created_at field
