+++ alembic/versions/reminder/699aad496719_add_new_table_reminder.py
@@ -1,0 +1,27 @@
+"""Add new table Reminder
+
+Revision ID: 699aad496719
+Revises: 9fac51c807d2
+Create Date: 2025-01-28 12:13:00.150533
+
+"""
+
+from typing import Sequence, Union
+
+# revision identifiers, used by Alembic.
+revision: str = "699aad496719"
+down_revision: Union[str, None] = "f69ce40494e3"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
