--- alembic/versions/reminder/9e82bf604fea_add_new_table_reminder.py
@@ -1,38 +1,0 @@
-"""Add new table Reminder
-
-Revision ID: 9e82bf604fea
-Revises: 03b93784fb26
-Create Date: 2025-01-23 11:44:11.718880
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "9e82bf604fea"
-down_revision: Union[str, None] = "f69ce40494e3"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "reminder",
-        sa.Column("id", sa.String(length=36), nullable=False),
-        sa.Column("user_id", sa.String(length=36), nullable=False),
-        sa.Column("title", sa.String(), nullable=False),
-        sa.Column("content", sa.String(), nullable=False),
-        sa.Column("creation_date", sa.DateTime(), nullable=False),
-        sa.Column("remind_date", sa.DateTime(), nullable=False),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("updated_at", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"]),
-        sa.PrimaryKeyConstraint("id"),
-    )
-
-
-def downgrade() -> None:
-    op.drop_table("reminder")
